Implement TC1300: RELATION-GROUP-TYPE
Also extract common code of Attribute Definition and Attribute Value parsing.

Closes #107
