+++ reqif/models/reqif_req_if_content.py
@@ -1,6 +1,8 @@
 from typing import List, Optional, Union
 
 from reqif.helpers.debug import auto_described
+from reqif.models.reqif_relation_group import ReqIFRelationGroup
+from reqif.models.reqif_relation_group_type import ReqIFRelationGroupType
 from reqif.models.reqif_spec_object import ReqIFSpecObject
 from reqif.models.reqif_spec_object_type import ReqIFSpecObjectType
 from reqif.models.reqif_spec_relation import ReqIFSpecRelation
@@ -20,13 +22,14 @@ class ReqIFReqIFContent:
                     ReqIFSpecObjectType,
                     ReqIFSpecRelationType,
                     ReqIFSpecificationType,
+                    ReqIFRelationGroupType,
                 ]
             ]
         ] = None,
         spec_objects: Optional[List[ReqIFSpecObject]] = None,
         spec_relations: Optional[List[ReqIFSpecRelation]] = None,
         specifications: Optional[List[ReqIFSpecification]] = None,
-        spec_relation_groups: Optional[List] = None,
+        spec_relation_groups: Optional[List[ReqIFRelationGroup]] = None,
     ):
         self.data_types: Optional[List] = data_types
         self.spec_types: Optional[
@@ -35,10 +38,13 @@ class ReqIFReqIFContent:
                     ReqIFSpecObjectType,
                     ReqIFSpecRelationType,
                     ReqIFSpecificationType,
+                    ReqIFRelationGroupType,
                 ]
             ]
         ] = spec_types
         self.spec_objects: Optional[List[ReqIFSpecObject]] = spec_objects
         self.spec_relations: Optional[List[ReqIFSpecRelation]] = spec_relations
         self.specifications: Optional[List[ReqIFSpecification]] = specifications
-        self.spec_relation_groups: Optional[List] = spec_relation_groups
+        self.spec_relation_groups: Optional[
+            List[ReqIFRelationGroup]
+        ] = spec_relation_groups
