+++ reqif/parsers/spec_relation_parser.py
@@ -6,7 +6,7 @@ from reqif.models.reqif_spec_relation import (
     ReqIFSpecRelation,
 )
 from reqif.models.reqif_types import SpecObjectAttributeType
-from reqif.parsers.spec_object_parser import ATTRIBUTE_STRING_TEMPLATE
+from reqif.parsers.attribute_value_parser import AttributeValueParser
 
 
 class SpecRelationParser:
@@ -37,6 +37,10 @@ class SpecRelationParser:
             attributes["LAST-CHANGE"] if "LAST-CHANGE" in attributes else None
         )
 
+        long_name: Optional[str] = (
+            attributes["LONG-NAME"] if "LONG-NAME" in attributes else None
+        )
+
         relation_type_ref = (
             xml_spec_relation.find("TYPE").find("SPEC-RELATION-TYPE-REF").text
         )
@@ -52,22 +56,32 @@ class SpecRelationParser:
         values_attribute: Optional[SpecObjectAttribute] = None
         xml_values = xml_spec_relation.find("VALUES")
         if xml_values is not None:
-            xml_string_attribute = xml_values.find("ATTRIBUTE-VALUE-STRING")
-            assert xml_string_attribute is not None
-            attribute_value = xml_string_attribute.attrib["THE-VALUE"]
-            definition_ref = xml_string_attribute[0][0].text
-            values_attribute = SpecObjectAttribute(
-                xml_node=xml_string_attribute,
-                attribute_type=SpecObjectAttributeType.STRING,
-                definition_ref=definition_ref,
-                value=attribute_value,
-            )
+            for xml_value in xml_values:
+                if xml_value.tag == "ATTRIBUTE-VALUE-STRING":
+                    attribute_value = xml_value.attrib["THE-VALUE"]
+                    definition_ref = xml_value[0][0].text
+                    values_attribute = SpecObjectAttribute(
+                        xml_node=xml_value,
+                        attribute_type=SpecObjectAttributeType.STRING,
+                        definition_ref=definition_ref,
+                        value=attribute_value,
+                    )
+                elif xml_value.tag == "ATTRIBUTE-VALUE-XHTML":
+                    values_attribute = (
+                        AttributeValueParser.parse_xhtml_attribute_value(
+                            xml_value
+                        )
+                    )
+                else:
+                    raise NotImplementedError
+                break
 
         spec_relation = ReqIFSpecRelation(
             xml_node=xml_spec_relation,
             description=description,
             identifier=identifier,
             last_change=last_change,
+            long_name=long_name,
             relation_type_ref=relation_type_ref,
             source=spec_relation_source,
             target=spec_relation_target,
@@ -83,6 +97,8 @@ class SpecRelationParser:
         output += f' IDENTIFIER="{spec_relation.identifier}"'
         if spec_relation.last_change is not None:
             output += f' LAST-CHANGE="{spec_relation.last_change}"'
+        if spec_relation.long_name is not None:
+            output += f' LONG-NAME="{spec_relation.long_name}"'
         output += ">\n"
 
         children_tags: List[str]
@@ -106,14 +122,10 @@ class SpecRelationParser:
                     spec_relation
                 )
             elif tag == "VALUES":
-                values_attribute = spec_relation.values_attribute
-                if values_attribute is not None:
-                    output += "          <VALUES>\n"
-                    output += ATTRIBUTE_STRING_TEMPLATE.format(
-                        definition_ref=values_attribute.definition_ref,
-                        value=values_attribute.value,
+                if spec_relation.values_attribute is not None:
+                    output += AttributeValueParser.unparse_attribute_values(
+                        [spec_relation.values_attribute]
                     )
-                    output += "          </VALUES>\n"
             else:
                 raise NotImplementedError(tag)
