+++ reqif/unparser.py
@@ -1,13 +1,18 @@
 from typing import List
 
 from reqif.models.reqif_namespace_info import ReqIFNamespaceInfo
+from reqif.models.reqif_relation_group_type import ReqIFRelationGroupType
 from reqif.models.reqif_spec_object_type import ReqIFSpecObjectType
 from reqif.models.reqif_spec_relation_type import ReqIFSpecRelationType
 from reqif.models.reqif_specification_type import ReqIFSpecificationType
 from reqif.parsers.data_type_parser import DataTypeParser
 from reqif.parsers.header_parser import ReqIFHeaderParser
+from reqif.parsers.relation_group_parser import ReqIFRelationGroupParser
 from reqif.parsers.spec_object_parser import SpecObjectParser
 from reqif.parsers.spec_relation_parser import SpecRelationParser
+from reqif.parsers.spec_types.relation_group_type_parser import (
+    RelationGroupTypeParser,
+)
 from reqif.parsers.spec_types.spec_object_type_parser import (
     SpecObjectTypeParser,
 )
@@ -60,7 +65,10 @@ class ReqIFUnparser:
                             reqif_xml_output += SpecificationTypeParser.unparse(
                                 spec_type
                             )
-
+                        elif isinstance(spec_type, ReqIFRelationGroupType):
+                            reqif_xml_output += RelationGroupTypeParser.unparse(
+                                spec_type
+                            )
                     reqif_xml_output += "      </SPEC-TYPES>\n"
 
                 if reqif_content.spec_objects is not None:
@@ -95,6 +103,14 @@ class ReqIFUnparser:
 
                 if reqif_content.spec_relation_groups is not None:
                     reqif_xml_output += "      <SPEC-RELATION-GROUPS>\n"
+
+                    for (
+                        spec_relation_group
+                    ) in reqif_content.spec_relation_groups:
+                        reqif_xml_output += ReqIFRelationGroupParser.unparse(
+                            spec_relation_group
+                        )
+
                     reqif_xml_output += "      </SPEC-RELATION-GROUPS>\n"
 
                 reqif_xml_output += "    </REQ-IF-CONTENT>\n"
