+++ tests/integration/examples/ci.eclipse.org/TC1300_E0000_S10_Reference_20210122_1256_jenkins/test.itest
@@ -1,4 +1,3 @@
-UNSUPPORTED: true
 RUN: mkdir -p %S/output
 RUN: %reqif passthrough %S/sample.reqif %S/output/sample.reqif
 RUN: diff %S/sample.reqif %S/output/sample.reqif
